requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
mediapipe>=0.10.0
dlib>=19.24.0
face-recognition>=1.3.0
librosa>=0.10.0
soundfile>=0.12.0
transformers>=4.30.0
tensorboard>=2.13.0
pillow>=10.0.0
pandas>=2.0.0 